new_message: Move get_sent update out of loop
Updating the get_sent query was previously done inside the loop over
inbox_rels, which meant both that it wasn't updated if inbox_rels was
empty, and that it was (unnecessarily) updated multiple times if there
are multiple rels.

This also moves the CachedQueryMutator instantiation outside the loop,
so we only use one instance if there are multiple rels.