Add flooring when displaying user karma totals
A floor of 1 total link karma already existed, this adds a floor to
comment karma as well. It also starts applying them everywhere,
including in the subreddit-specific karma breakdown.

In addition, admins are now exempt from the flooring and can see the
actual karma totals for users in all locations.