trunc_string: split on word boundaries
This makes it so that instead of seeing text like this:

    This is a long tit...

we'll get

    This is a long...

when truncating "This is a long title".  Usually this is just a minor
improvement, but occasionally odd truncation can [lead to hilarious
results][0].  And, y'know, I don't like hilarity.

We currently only use `trunc_string` in a few places, all related to the
displayed titles or descriptions in Facebook OpenGraph notes or Twitter Cards,
so the impact isn't too terribly great.  We are now including the suffix in the
length count, so that might actually fix a bug.

We don't use the `length` parameter by keyword anywhere, so changing it isn't a
problem.

[0]: http://blogs.hbr.org/2012/07/i-wont-hire-people-who-use-poo/