Pad submit buttons
I noticed that in Chrome, some of our buttons had no padding at all, which made
them look terrible.  Others, however, were fine.  This turned out to be a
different between `<input type="submit">` and `<button>` elements.  The easiest
place to see this was comparing the save buttons on `/prefs` (ugly) and any of
the other preferences tabs (pretty).

In `reddit.less`, we reset the margin and padding of a whole bunch of elements
to be 0.  `<input>`s are in this list, although `<button>`s are not.\*  Thus,
no padding.

Now we set the padding `<button>`s and select types of `<input>`s.  I chose the
rules from what Chrome automatically applies, since that seemed like a
reasonable default.

Firefox notably ignored the `padding: 0` before, and continues to ignore
whatever I set, with the result that it looked good before and still looks
good now.  Whatever.

----

\* I decided not to add a reset of button margin and padding.  Padding we now
set again to something useful, per this commit, but margin could have effects
across the site that I didn't want to track down.