Add option for sending orangereds as emails
Some users visit fairly infrequently, or visit without being logged-in, and so
[don't notice messages they've received][0].  We need some better data on this,
and it's hard to separate these people out from alts (particularly novelty
accounts).  However, [anecdotal evidence][1] and conversations shows that there
is *some* demand for this feature.

There are a lot more things to do to get this ready for the public.  However,
we just want to get a quick prototype out so that we can start using it
internally and see where it falls apart.

[0]: https://www.reddit.com/r/redditdev/comments/2f12s3/email_notifications/
[1]: https://ifttt.com/recipes/94391-reddit-inbox-to-email-inbox