Make the gilded message repliable
The gold message is sent to the gildee with the gilder as the author,
but the username is hidden through the user of the display_author attr.
display_author and display_to fields have been added to keep track of
the labels to be displayed. If there is a display_to field, then it is
a reply to an anonymous message (one that started with a display_author).
A warning will be appended to messages that have display_to saying that
their identity will be revealed if they reply.