Feature flags: allow filtering by subdomain
We'd like to have an easy way to push out changes to a subset of users before
going to everyone.  Like with the previous change to add subreddit filtering,
toggling features by subdomain gives us that ability fairly cheaply.

(There is also a more sophisticated [betamode plugin], but it's designed for
more heavyweight needs and requires more setup for each test.)

[betamode plugin]: https://github.com/reddit/reddit-plugin-betamode