Determine if a comment is gildable from can_save
Previously this was being done with can_reply, but that can make
archived comments, comments in subreddits the user is banned from, etc.
ungildable. Now that comment-saving is available to all users (and not
just ones with gold like it was at the time), can_save is a better
indicator to use.