Delay email sends
When someone spams all over the place, we usually catch it before people check
their notifications.  If we send them an email, though, we lose the ability to
take it back.

To address this, we're going to hold on to notifications for a bit before
sending them.  I don't know how this particular approach will scale, so we'll
need to test that.

It also unfortunately makes things a lot more difficult to test. :(