vote: Add {Link,Comment}VoteDetailsByDay CFs.
We want to ditch the postgres-based vote relations as they are not being
used for much right now and make a large ops burden as they use a
massive amount of space.

These column families should take care of one of the last unserved use
cases for votes: day-based data dumps: the about page shows a count of
votes day-to-day, and some awards use yesterday's votes to determine
winners.

I've chosen to use Pacific time because that's when award cron jobs run
and is generally still how we think about "day"s since we're here in SF.