Language modal: fix on login page
In reddit/reddit@c4f5b6c2, we accidentally broke the language-selector modal
that appears in the top-right when logged-out (when logged-in language is
changed from your preferences).  The `<div>` that gets popped up snuck up
inside one of the many if-statements in that section, which resulted in it not
getting added to the html on pages where we want to hide the login modal link
(currently only login pages).  This restores it, so you can once again change
your language while on /login.