Consistencize account text
To avoid debate over "login" the noun and "log in" the verb, we're making a
move towards using "sign in" instead.  The first part of this was done in the
login rework in reddit/reddit@c4f5b6c2.

This commit continues to make our verbage consistent on "sign in" and "create
an account", avoiding use of "log in" and "register".  Specifically, these
should all be consistent:

* Upper-right modal-popping text
* Headers on modal and /login
* Buttons on modal and /login
* Page title on /login

I didn't change the text in the old login flow, because that feature flag seems
like it's going to be pruned soon.