markdown: Fix more issues with new markdown styles.
+ fixes accidental change to sidebar styles in live styles
+ restores ability to do intentionally spaced out lists
+ fixes compressed comments on user pages

Markdown styles remove the top margin from the first element and the bottom margin
from the last element, so that these can be effectively controlled on the container
element.  The userpage lacked any styling for this, so the first line of a comment
appeared squished up against the header line.