Inbox counts: Remove live alter of unread count when mark-on-open is off
This is a good interaction in theory, but gets pretty hairy to test for
in practice due to modmail being a subset of messages as well as
subreddit messages not being restricted to modmail. This issue is
blocking release, so remove the buggy feature for rollout to be addressed
at a later date.