Websockets: Don't send stats to graphite on enqueue.
This is currently causing a separate stat to be added to graphite for
each live thread that gets created. This is dangerous because it causes
a lot of growth of carbon files on disk which could represent a DoS
opportunity. These statistics are not terribly useful to know as they
directly correlate to actions in the live thread which can be seen in
other ways.