selfserve: Return to first page when changing sort on sponsor/promoted/ listing.
Currently, if you navigate to the second page of links then change the sort type, you'll
get no results back.  This is because it is carrying over the 'after' query param.  This
drops the query params for all of the sort options _except_ the one you are currently
on, so that changing it will always kick you back to page 1.