Orangered emails: include sender's name
@umbrae pointed out that it's nice to know who a message is from, even if you
have to go to the message on the web to reply.

It's possible at some point in time we will need something else from the sender
than just their username.  However, sending the id and causing a lookup seems
unnecessary for a vague possibility - and we should be able to make a
transition in the future pretty easily if necessary.