normalized_hot: Remove caching of normalized values.
Hypothesis: this sgm is doing exactly two things:
  * caching the computation of ehot values
  * moving read-pressure (for CachedResults) from the permamemcaches to
    the main memcaches

The computation of ehot values consists of dividing the hot score for
each item by the highest hot value in the query.  Given that the are at
most 1000 links per query and up to 100 subreddits in a listing, this
means at most 100,000 division ops which are still quite fast. timeit
shows this computation in worst case takes about 5ms which is likely
faster than a network roundtrip, especially during congestion. There
might be some app CPU savings involved here that we should definitely
monitor.

Moving read-pressure from the permamemcaches to the main memcaches is
slightly more concerning, but we can determine quickly if the
permamemcaches will be maxed out by the transition.

The goal here is to remove the sgm step and as a result reduce network
traffic to the main memcaches.