Comment embeds: fix subreddit link
At the bottom of the comment embeds is a link to the subreddit from which the
comment was extracted.  This link has had an extra domain appended, so we'd end
up with URLs like `http://www.redditmedia.com/www.reddit.com/r/worldnews`.

The problem stemmed from the fact that `get_domain()` returns a protocol-less
URL (`www.reddit.com/r/worldnews`).  Thus, `add_sr()` (which it ends up being
passed into) assumes it has been handed a relative URL and appends a domain.

Instead of getting the entire domain, we now just get the subreddit path and
let `add_sr()` handle the domain (and protocol) later.